Split up remaining files
