+++ course_flow/static/course_flow/js/react/src/Components/Views/OutcomeTableView/index.js
@@ -1,3 +1,4 @@
 export { default as OutcomeTableView } from './OutcomeTableView.js'
 export { default as OutcomeBase } from './OutcomeBase.js'
 export { OutcomeUnconnected } from './Outcome.js'
+export { default as OutcomeLegend } from './OutcomeLegend.js'
