+++ course_flow/static/course_flow/js/react/src/Components/Views/WorkflowView/index.js
@@ -2,3 +2,5 @@ export { default as WorkflowView } from './WorkflowView.js'
 export { NodeWeekUnconnected as NodeWeekUnconnected } from './NodeWeek.js'
 export { WeekUnconnected as WeekUnconnected } from './Week.js'
 export { WeekWorkflowUnconnected as WeekWorkflowUnconnected } from './WeekWorkflow.js'
+export { default as OutcomeNode } from './OutcomeNode.js'
+export { WorkflowLegendUnconnected as WorkflowLegendUnconnected } from './WorkflowLegend.js'
