+++ course_flow/static/course_flow/js/react/src/Components/components/CommonComponents/LegendLine.js
@@ -1,0 +1,21 @@
+import * as React from 'react'
+
+class LegendLine extends React.Component {
+  /*******************************************************
+   * RENDER
+   *******************************************************/
+  render() {
+    let icon
+    if (this.props.icon)
+      icon = <img src={config.icon_path + this.props.icon + '.svg'} />
+    else icon = <div className={this.props.divclass}>{this.props.div}</div>
+    return (
+      <div className="legend-line">
+        {icon}
+        <div>{this.props.text}</div>
+      </div>
+    )
+  }
+}
+
+export default LegendLine
